pyproject.toml
[tool.poetry]
name = "continua"
version = "0.1.0"
description = ""
authors = ["Bilyana Tzolova <bmtzolova@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.1"
boto3 = "^1.35.0"
python-dotenv = "^1.0.1"
ollama = "^0.3.1"
numpy = "^2.1.0"
anthropic = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
